backend/requirements.txt
fastapi
uvicorn
openpyxl
python-multipart
pywebview
